datalake/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
delta-spark==3.2.1
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.54.3
packaging==24.2
pandas==2.2.3
plotly==5.24.1
py4j==0.10.9.7
pydantic==2.9.2
pydantic_core==2.23.4
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.0.6
zipp==3.20.2
